+++ src/lib/components/ExplorerLink.tsx
@@ -1,12 +1,11 @@
-import type { BoxProps } from "@chakra-ui/react";
+import type { BoxProps, TextProps } from "@chakra-ui/react";
 import { Box, Text } from "@chakra-ui/react";
 import { useWallet } from "@cosmos-kit/react";
 
 import {
   getExplorerBlockUrl,
-  getExplorerTxUrl,
-  getExplorerValidatorUrl,
   getProposalUrl,
+  getExplorerValidatorUrl,
 } from "lib/app-fns/explorer";
 import type { AddressReturnType } from "lib/app-provider";
 import { AmpTrackMintscan } from "lib/services/amplitude";
@@ -30,6 +29,8 @@ interface ExplorerLinkProps extends BoxProps {
   isReadOnly?: boolean;
   textFormat?: "truncate" | "ellipsis" | "normal";
   maxWidth?: string;
+  textVariant?: TextProps["variant"];
+  ampCopierSection?: string;
 }
 
 const getNavigationUrl = (
@@ -40,7 +41,7 @@ const getNavigationUrl = (
   let url = "";
   switch (type) {
     case "tx_hash":
-      url = getExplorerTxUrl(currentChainName);
+      url = "/tx";
       break;
     case "contract_address":
       url = "/contract";
@@ -92,6 +93,7 @@ const LinkRender = ({
   textValue,
   isEllipsis,
   maxWidth,
+  textVariant,
 }: {
   type: string;
   isInternal: boolean;
@@ -99,10 +101,11 @@ const LinkRender = ({
   textValue: string;
   isEllipsis: boolean;
   maxWidth: ExplorerLinkProps["maxWidth"];
+  textVariant: TextProps["variant"];
 }) => {
   const textElement = (
     <Text
-      variant="body2"
+      variant={textVariant}
       color="lilac.main"
       transition="all .25s ease-in-out"
       _hover={{ color: "lilac.light" }}
@@ -141,14 +144,17 @@ export const ExplorerLink = ({
   showCopyOnHover = false,
   isReadOnly = false,
   textFormat = "truncate",
-  maxWidth = "150px",
+  maxWidth = "160px",
+  textVariant = "body2",
+  ampCopierSection,
   ...componentProps
 }: ExplorerLinkProps) => {
   const { address, currentChainName } = useWallet();
   const isInternal =
     type === "code_id" ||
     type === "contract_address" ||
-    type === "user_address";
+    type === "user_address" ||
+    type === "tx_hash";
 
   const [hrefLink, textValue] = [
     getNavigationUrl(type, currentChainName, copyValue || value),
@@ -182,6 +188,7 @@ export const ExplorerLink = ({
             textValue={textValue}
             isEllipsis={textFormat === "ellipsis"}
             maxWidth={maxWidth}
+            textVariant={textVariant}
           />
           <Copier
             type={type}
@@ -189,6 +196,7 @@ export const ExplorerLink = ({
             copyLabel={copyValue ? `${getCopyLabel(type)} Copied!` : undefined}
             display={showCopyOnHover ? "none" : "block"}
             ml="8px"
+            amptrackSection={ampCopierSection}
           />
         </>
       )}