+++ src/lib/components/TokenCard.tsx
@@ -1,3 +1,4 @@
+import type { FlexProps } from "@chakra-ui/react";
 import { Badge, Flex, Image, Text, Tooltip } from "@chakra-ui/react";
 
 import { NAToken } from "lib/icon";
@@ -10,11 +11,16 @@ import {
 
 import { Copier } from "./copy";
 
-interface TokenCardProps {
+interface TokenCardProps extends FlexProps {
   userBalance: BalanceWithAssetInfo;
+  amptrackSection?: string;
 }
 
-export const TokenCard = ({ userBalance }: TokenCardProps) => {
+export const TokenCard = ({
+  userBalance,
+  amptrackSection,
+  ...cardProps
+}: TokenCardProps) => {
   const { symbol, price, amount, precision, id } = userBalance.balance;
 
   return (
@@ -35,6 +41,7 @@ export const TokenCard = ({ userBalance }: TokenCardProps) => {
         p={3}
         background="pebble.900"
         borderRadius="8px"
+        {...cardProps}
       >
         <Flex
           gap={1}
@@ -67,6 +74,7 @@ export const TokenCard = ({ userBalance }: TokenCardProps) => {
             copyLabel="Token ID Copied!"
             display="none"
             ml="1px"
+            amptrackSection={amptrackSection}
           />
         </Flex>