+++ src/lib/components/copy/Copier.tsx
@@ -11,6 +11,7 @@ interface CopierProps {
   copyLabel?: string;
   display?: LayoutProps["display"];
   ml?: string;
+  amptrackSection?: string;
 }
 
 export const Copier = ({
@@ -19,6 +20,7 @@ export const Copier = ({
   copyLabel,
   display = "block",
   ml = "8px",
+  amptrackSection,
 }: CopierProps) => (
   <CopyTemplate
     value={value}
@@ -30,7 +32,7 @@ export const Copier = ({
         display={display}
         cursor="pointer"
         marginLeft={ml}
-        onClick={() => AmpTrackCopier(type)}
+        onClick={() => AmpTrackCopier(amptrackSection, type)}
         name="copy"
         boxSize="12px"
       />