+++ src/lib/hooks/index.ts
@@ -6,3 +6,4 @@ export * from "./useRedo";
 export * from "./useSingleMessageProps";
 export * from "./useToast";
 export * from "./useUserKey";
+export * from "./useOpenTxTab";