+++ src/lib/services/assetService.ts
@@ -1,31 +1,37 @@
 import { useWallet } from "@cosmos-kit/react";
-import type { UseQueryResult } from "@tanstack/react-query";
 import { useQuery } from "@tanstack/react-query";
-import { useCallback } from "react";
 
 import { getAssetInfos } from "lib/services/asset";
-import type { AssetInfo } from "lib/types";
+import type { AssetInfo, Option } from "lib/types";
 
-export const useAssetInfos = (): UseQueryResult<Record<string, AssetInfo>> => {
+export type AssetInfosOpt = Option<{ [key: string]: AssetInfo }>;
+
+export const useAssetInfos = (): {
+  assetInfos: AssetInfosOpt;
+  isLoading: boolean;
+} => {
   const { currentChainRecord } = useWallet();
 
-  const queryFn = useCallback(
-    async () =>
-      getAssetInfos(
-        currentChainRecord?.name,
-        currentChainRecord?.chain.chain_id
-      ),
-    [currentChainRecord?.chain.chain_id, currentChainRecord?.name]
-  );
-
-  return useQuery(
+  const { data: assets, isLoading } = useQuery(
     [
       "query",
       "assetInfos",
       currentChainRecord?.name,
       currentChainRecord?.chain.chain_id,
     ],
-    queryFn,
+    async () =>
+      getAssetInfos(
+        currentChainRecord?.name,
+        currentChainRecord?.chain.chain_id
+      ),
     { enabled: !!currentChainRecord }
   );
+
+  return {
+    assetInfos: assets?.reduce(
+      (acc, asset) => ({ ...acc, [asset.id]: asset }),
+      {}
+    ),
+    isLoading,
+  };
 };