+++ src/lib/services/txQuery/useTxQuery.ts
@@ -207,6 +207,7 @@ export const useTxQuery = (
       indexerGraphClient,
     ],
     queryFn,
+    enabled: !!userAddress,
   });
 };
 
@@ -257,5 +258,6 @@ export const useTxQueryCount = (
       indexerGraphClient,
     ],
     queryFn,
+    enabled: !!userAddress,
   });
 };