+++ src/lib/types/tx/common.ts
@@ -1,0 +1,5 @@
+import type { StdFee } from "@cosmjs/stargate";
+
+export interface Fee extends Omit<StdFee, "gas"> {
+  gas_limit: string;
+}